implement refresh accessToken

+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/user/function/FindUserById.kt
@@ -7,8 +7,8 @@ import java.util.*
 
 class FindUserById(
     private val userRepository: UserRepository
-): Function<Int, User> {
-    override fun apply(userId: Int): User {
+): Function<Long, User> {
+    override fun apply(userId: Long): User {
 
         return userRepository.findById(userId).get()
     }
